[BE] fix: 자신의 세션 만료시킬 때 제대로 만료되지 않는 문제 수정
- sid 비교해서 로그아웃 호출

+++ WEB/BE/controllers/web/log/index.js
@@ -5,9 +5,15 @@ const redis = require('@models/redis');
 const logController = {
   async delSession(req, res, next) {
     const { sid } = req.body;
+
+    if (sid === req.session.id) {
+      return next();
+    }
+
     const client = redis.client();
     client.del(`session:${sid}`);
     await logService.update({ sid, isLoggedOut: true });
+
     res.json({ result: true });
   },
   async get(req, res) {
